requirements.txt
fastmcp==2.14.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.14.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.3
